pyproject.toml
[project]
name = "experimental-llama.cpp"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    { name = "David Ng", email = "david.ng.dev@gmail.com" },
]
dependencies = [
    "httpx>=0.27.2",
    "starlette>=0.41.2",
    "uvicorn>=0.32.0",
    "llama-cpp-python>=0.3.1",
    "huggingface-hub[cli]>=0.26.2",
]
requires-python = "==3.12.*"
readme = "README.md"
license = { text = "MIT" }

[tool.pdm]
distribution = false

[tool.pdm.dev-dependencies]
dev = [
    "pytest>=8.3.3",
]

[tool.pytest.ini_options]
pythonpath = [
    "src"
]